requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements-dev.in
#
alembic==1.16.4
    # via -r requirements-dev.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
autoflake==2.3.1
    # via -r requirements-dev.in
bandit==1.8.6
    # via -r requirements-dev.in
black==25.1.0
    # via -r requirements-dev.in
build==1.3.0
    # via
    #   -r requirements-dev.in
    #   pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via
    #   black
    #   pip-tools
    #   uvicorn
fastapi==0.116.1
    # via -r requirements.in
flake8==7.3.0
    # via -r requirements-dev.in
greenlet==3.2.4
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements-dev.in
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.17.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   mypy
pip-tools==7.5.0
    # via -r requirements-dev.in
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
pycodestyle==2.14.0
    # via flake8
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.4.0
    # via
    #   autoflake
    #   flake8
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
pytest-asyncio==1.1.0
    # via -r requirements-dev.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   bandit
    #   uvicorn
rich==14.1.0
    # via
    #   -r requirements.in
    #   bandit
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.47.3
    # via fastapi
stevedore==5.5.0
    # via bandit
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via
    #   -r requirements-dev.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
